chore: refactor component imports and add HeroSectionPlaceHolder; group footer components
