+++ src/routes/-components/footer-section/FooterCTA.tsx
@@ -1,7 +1,7 @@
 import { Link } from "@tanstack/react-router";
-import rocketIcon from "../../assets/rocket.png";
-import robot from "../../assets/robot-assistant.png";
-import ellipse from "../../assets/ellipse.png";
+import rocketIcon from "@/assets/rocket.png";
+import robot from "@/assets/robot-assistant.png";
+import ellipse from "@/assets/ellipse.png";
 export const FooterCTA = () => {
   return (
     <div
